requirements.txt
flask==2.3.3
crawl4ai==0.4.24
aiohttp==3.11.12
beautifulsoup4==4.12.2
weaviate-client
transformers
torch 